Bring server values into template using cookies
Simply excruciating. Got it to work but this is a prime example
of "Angular is not the right fit for what we want".

Since cookies are available on the server and the client, it's
fine for now, and we don't need any further weird data caching.

Greatly helped by this stackoverflow answer:
- https://stackoverflow.com/questions/45750187/using-cookies-in-universal-rendering-using-angular-cli/46266296#46266296

Note that I added "trim()" to `parseCookie` in the cookie service
because otherwise we would get keys with leading spaces.

Other references:
- https://github.com/angular/universal/issues/709
- https://github.com/angular/universal-starter/issues/271
- https://github.com/angular/universal-starter/issues/373

VERY FUN.